requirements.txt
click
PyYAML
rich
textual
# fzf должен быть установлен отдельно в системе